integration: diva (#352)
Creates a Diva Beacon Oracle contract that can prove:

Any of the fields of a validator.

The balance of a validator using the balances container in BeaconState.

A deposit from a BLSPubkey using the deposits container in BeaconBlock.

Note:

Lodestar currently has a gindex issue for fields in the Validator container. As such, we need to calculate the gindices of Validator fields manually.

GitOrigin-RevId: 26db3160591809821b4785bce105ac980c0db6bd